BC when bleedingEdge is disabled
