+++ src/Testing/RuleTestCase.php
@@ -98,6 +98,7 @@ abstract class RuleTestCase extends PHPStanTestCase
 				[],
 				true,
 				$this->shouldTreatPhpDocTypesAsCertain(),
+				self::getContainer()->getParameter('featureToggles')['bleedingEdge'],
 			);
 			$fileAnalyser = new FileAnalyser(
 				$this->createScopeFactory($reflectionProvider, $typeSpecifier),
