+++ tests/PHPStan/Rules/Exceptions/disable-bleeding-edge.neon
@@ -1,0 +1,3 @@
+parameters:
+	featureToggles:
+		bleedingEdge: false
